Make checkbox labels clickable
HTML provides two ways to associate a label with its associated
checkbox:
1. specify the checkbox's id in the label's `for` attribute
2. nest the checkbox inside the <label> element.

Unfortunately, VulcanJS uses neither of these methods, so we lose some
advantages from the browser understanding the association. When a
label is associated with its checkbox:
1. Clicking the label toggles the checkbox
2. Screen readers describe the association better

Ideally, I'd alter VulcanJS form code to make it associate the labels
correctly, but it seemed difficult, so instead I just replicate the
first advantage, labels as checkbox click targets, with a JavaScript
event listener, in clickableCheckboxLabels.ts.

This applies only to checkboxes in forms generated by VulcanJS from
collection schemas. Other checkboxes, for which we write the code
ourselves, should use one of the two HTML ways to associate with labels
correctly, so this commit also changes some of those to associate
them the usual HTML way.